athena: postgres12 docker image (#36)
This copies the existing `postgres11` settings and `Dockerfile` to a new `postgres12` folder, in order to enable using `postgres12` on CircleCI for `build_and_test` and for local development. We currently use PostgreSQL 12.3 in our Heroku add-on for `athena_logic` in staging, so this would let our development and builds match our deployments. For comms and DnD, the staging environment is still using 11.8, but we should update their DBs to 12.3 to be consistent.

The only changes between this and `postgres11` currently are renaming `POSTGRES11_DATABASE_NAMES` to `POSTGRES12_DATABASE_NAMES` and setting the `FROM` in the `Dockerfile` to `12.3-alpine`.