resware-mssql: Remove a temporary echo statement.
This reverts StatesTitle/dockerhub#55. It was a temporary echo statement used to debug a problem. @knoight said "i think you can revert it - we’ve learned what the problem is - once we have a potential fix, we should be able to test it without that debug statement, because if it fixes the problem, it will just work."

Maybe a little more info in [this Slack thread](https://statestitle.slack.com/archives/C010MJ7SZN2/p1626974904011200).