create a customized circleci image to use instead of heroku (#74)
The goal of this change is to replace the `statestitle/underwriter` image that we use in our `config.yml` file in the `underwriter` monorepo with `statestitle/circleci`. The `Dockerfile` for `statestitle/circleci` hasn't been updated in a while, and all of our existing images still use `heroku` installers or base images - those images are generally running Ubuntu 16.x, which is no longer receiving standard support; the updated `cimg/python:3.7.13` image is running `cimg/base:2022.01` which is based on Ubuntu 20.x, which is [LTS until 2025](https://wiki.ubuntu.com/Releases).

Additionally we don't need to do custom builds of the `thrift` binaries for python; the `thrift-compiler` package on Ubuntu 20.x is now at 0.13.0.

For Node and Yarn, there are `cimg/python:3.7-node` images that pull in the latest LTS (which would be 16.x currently), but I didn't want to cause breaks when that suddenly changes to 18.x in the fall. Instead I've pinned those versions here like they were in the `underwriter` image and copied the commands that the `cimg` Dockerfiles were using.

For testing, I created an image locally and pushed it to dockerhub as `statestitle/circleci:cplummer-test` then created a branch in the `underwriter` repo with the new image - see https://github.com/StatesTitle/underwriter/pull/15020 for the draft; all the CI tests passed. I also ran `docker run --rm -it --entrypoint bash statestitle/circleci:cplummer-test` locally and checked that the `thrift -version` command produced `0.13.0` with the same type of test for `node` and `yarn`.